+++ src/components/Label/Label.tsx
@@ -2,6 +2,7 @@ import styled, { css } from 'styled-components';
 import { ifProp } from 'styled-tools';
 import { applyStyleModifiers } from 'styled-components-modifiers';
 import { fontSizeExtraSmall, fontSizeSmall, fontSizeMedium, fontSizeLarge } from '../../modifiers';
+import { AnalyticsProps } from '../../analytics';
 
 const typographyModifiers = {
   fontSizeExtraSmall,
@@ -10,7 +11,7 @@ const typographyModifiers = {
   fontSizeLarge,
 };
 
-const Label = styled.label<{ disabled?: boolean }>`
+const Label = styled.label<{ disabled?: boolean } & AnalyticsProps>`
   font-size: ${({ theme }) => theme.font.sizes.medium};
   font-weight: ${({ theme }) => theme.font.weights.normal};
   color: ${({ theme }) => theme.palette.text.main};