+++ src/index.ts
@@ -42,3 +42,6 @@ export { default as ThemeWrapper } from './theme/ThemeWrapper';
 
 // colors
 export { default as Colors } from './theme/colors';
+
+// analytics
+export * from './analytics';